chore(setup):setup-postgres-and-sequelize
putting a line at the end of line.

[ready for review #182201659]
