+++ .sequelizerc
@@ -3,8 +3,8 @@ require('@babel/register') ;
 const { resolve } =require( 'path');
 
 module.exports= {
-  'config': resolve('src', 'config','config.js'),
-  'migrations-path': resolve('src', 'migrations'),
-  'models-path': resolve('src', 'models'),
-  'seeders-path': resolve('src', 'seeders')
+  'config': resolve('src','database', 'config','config.js'),
+  'migrations-path': resolve('src','database', 'migrations'),
+  'models-path': resolve('src','database', 'models'),
+  'seeders-path': resolve('src','database', 'seeders')
 }
