+++ src/database/config/config.js
@@ -27,3 +27,4 @@ export default {
   dialect: 'postgres',
   },
 }
+
