+++ src/database/models/index.js
@@ -33,3 +33,4 @@ db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
 export default db;
+
