+++ src/database/seeders/20220523193745-demo-user.js
@@ -12,3 +12,4 @@ export async function up(queryInterface, Sequelize) {
 export async function down(queryInterface, Sequelize) {
   await queryInterface.bulkDelete('users', null, {});
 }
+
