+++ src/index.js
@@ -1,5 +1,6 @@
-import app from './routes/index';
+import app from './app';
 import 'dotenv/config';
+
 const PORT = process.env.PORT || 5000;
 app.listen(PORT, () => {
   console.log(`app is listening on port ${PORT}`);
